requirements.txt
visdom
attrdict
PyYAML
pillow